refactor(mnemonic_enhancer): move mnemonic enhancement and database to _from data_prep to _mnemonic_enhancer module
