+++ src/_mnemonic_enhancer/mnemonic_ft.py
