+++ src/_mnemonic_enhancer/mnemonic_schemas.py
